requirements.txt
a-parser==3.0.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
html-sanitizer==1.9.2
idna==3.3
lxml==4.7.1
python-slugify==5.0.2
python-wordpress-xmlrpc==2.3
random-user-agent==1.0.1
requests==2.7.0
soupsieve==2.2.1
text-unidecode==1.3
urllib3==1.26.7
